fixed the linting issue & refactored tests
